Move common code into a separate package
This PR moves the common code currently in the iptables
package to a separate package, so that it can be shared
between the iptables and nftables implementations without
duplicating code.
