+++ cni/pkg/plugin/sidecar_iptables_linux.go
@@ -22,8 +22,8 @@ import (
 	"github.com/containernetworking/plugins/pkg/ns"
 
 	"istio.io/istio/pkg/log"
+	"istio.io/istio/tools/common/config"
 	"istio.io/istio/tools/istio-iptables/pkg/cmd"
-	"istio.io/istio/tools/istio-iptables/pkg/config"
 	"istio.io/istio/tools/istio-iptables/pkg/dependencies"
 )
