+++ tools/common/tproxy/run_linux.go
@@ -11,18 +11,19 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package capture
+package tproxy
 
 import (
 	"fmt"
 	"net"
 	"strconv"
+	"strings"
 
 	"github.com/vishvananda/netlink"
 	"golang.org/x/sys/unix"
 
 	"istio.io/istio/pkg/log"
-	"istio.io/istio/tools/istio-iptables/pkg/config"
+	"istio.io/istio/tools/common/config"
 )
 
 // configureTProxyRoutes configures ip firewall rules to enable TPROXY support.
@@ -124,4 +125,14 @@ func configureIPv6Addresses(cfg *config.Config) error {
 	}
 	log.Infof("Added ::6 address on the loopback iface")
 	return nil
+}
+
+func ignoreExists(err error) error {
+	if err == nil {
+		return nil
+	}
+	if strings.Contains(strings.ToLower(err.Error()), "file exists") {
+		return nil
+	}
+	return err
 }
