+++ tools/istio-iptables/pkg/builder/iptables_builder_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	testutil "istio.io/istio/pilot/test/util"
-	"istio.io/istio/tools/istio-iptables/pkg/config"
+	"istio.io/istio/tools/common/config"
 )
 
 func compareToGolden(t *testing.T, name string, actual string) {
